@@ -8,33 +8,33 @@ authors = ["László Andrási <mail@laszloandrasi.com>"]
 g2 = "gitential2.cli_v2:main"
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.10"
 invoke = "^1.4.1"
 pydantic = "^1.7.2"
-pygit2 = "^1.4.0"
+pygit2 = "^1.7.1"
 numpy = "^1.19.4"
 cryptography = "^3.2.1"
 click = "^7.1.2"
 tqdm = "^4.53.0"
 structlog = "^20.1.0"
-fastapi = "^0.62.0"
-uvicorn = "^0.12.3"
+fastapi = "^0.70.0"
+uvicorn = "^0.15.0"
 requests = "^2.25.0"
 websockets = "^8.1"
 sqlalchemy = "^1.3.22"
 psycopg2 = "^2.8.6"
 alembic = "^1.4.3"
-sqlalchemy-utils = "^0.36.8"
+sqlalchemy-utils = "^0.37.9"
 pathspec = "^0.8.1"
 pyyaml = "^5.3.1"
-authlib = "^0.15.3"
-httpx = "^0.16.1"
+authlib = "^0.15.5"
+httpx = "^0.21.1"
 itsdangerous = "^1.1.0"
 loginpass = "^0.5"
 mypy = "^0.800"
 celery = "^5.0.5"
 redis = "^3.5.3"
-ibis-framework = "^1.4.0"
+ibis-framework = "^2.0.0"
 rsa = "^4.7.2"
 billiard = "^3.6.3"
 jinja2 = "^2.11.3"
@@ -47,7 +47,7 @@ stripe = "^2.58.0"
 email-validator = "^1.1.3"
 
 [tool.poetry.dev-dependencies]
-pytest = "^5.2"
+pytest = "^6.2.5"
 black = {version = "^20.8b1", allow-prereleases = true}
 pylint = "^2.6.0"
 mypy = "^0.800"
